# Optimize roll by using foldl' instead of foldr