@kolmodin 2015-10-06
Looks great!

Unrelated to you change, but something I'd like to look at; I don't think the larger values of Integer gets tested in the QC tests. I think Integer needs something similar as we did for the QC tests for Numeric.Natural.