@kolmodin 2015-10-06
Ah, looks like the Integer instance already had tests that covered both paths.